requirements.txt
aiofiles==23.1.0
anyio==3.6.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.4
h11==0.14.0
html5tagger==1.3.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nodeenv==1.8.0
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
prisma==0.8.2
psutil==5.9.5
pydantic==1.10.8
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sanic==23.3.0
sanic-ext==23.3.0
sanic-routing==22.8.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.24.0
six==1.16.0
sniffio==1.3.0
sympy==1.12
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
tracerite==1.1.0
typing_extensions==4.6.1
tzdata==2023.3
ujson==5.7.0
ultralytics==8.0.109
urllib3==1.26.16
uvloop==0.17.0
websockets==11.0.3
